Process host I/O events before any foreign function invocation
This ensures that the readiness of host-backed file descriptions are
up-to-date before executing a foreign function on it.
